feat: use cheer modal 훅 생성 및 적용
