+++ components/molecules/cheer/index.ts
@@ -1,2 +1,3 @@
 export * from './cheer-bottom-sheet';
 export * from './cheer-item';
+export * from './cheer-progress';
