+++ features/record-detail/components/index.ts
@@ -1,5 +1,4 @@
 export * from './cheer-modal';
-export * from './cheer-progress';
 export * from './date-picker';
 export * from './edit-button';
 export * from './swim-description-item';
